+++ src/components/Forms/form-patient-login/form-patient-login.tsx
@@ -2,43 +2,48 @@ import React from "react";
 import {
   ContainerErrorText,
   ContainerFormPatientRecord,
-  ContainerIconInput,
-  ContainerInput,
   ContainerLogin,
-  Input,
   InputText,
 } from "./form-patient-login-style";
 import Icon from "react-native-vector-icons/Ionicons";
-import { Image, Text } from "react-native";
-import { ImageSource } from "react-native-vector-icons/Icon";
-import { ImageLogo } from "../../Buttons/buttonEmailSenha/buttonEmailSenha-style";
-import { InputLogin, PatientInput } from "../../../screen/login/login-style";
-import InputGeral from "../inputGeral/inputGeral";
+import { PatientInput } from "../../../screen/login/login-style";
+
 
 interface PatientRecordValues {
   placeholder: string;
   handleInputChange: any;
   state: any;
   err: any;
-  imageButton: ImageSource;
+  showPassword: boolean;
 }
 
-const FormPatientLoginRecord = ({ imageButton, placeholder,
-  handleInputChange, state, err }: PatientRecordValues) => {
+const FormPatientLoginRecord = ({ placeholder,
+  handleInputChange, state, err}: PatientRecordValues) => {
+  const [showPassword, setShowPassword] = React.useState(false);
+
+  const toggleShowPassword = () => {
+    setShowPassword(!showPassword);
+  };
 
   return (
     <ContainerFormPatientRecord>
       <ContainerLogin>
-        <ImageLogo>
-          <Image source={imageButton} />
-        </ImageLogo>
         <PatientInput
           onChangeText={handleInputChange}
           value={state}
           placeholder={placeholder}
           placeholderTextColor="#4c7abb"
-          secureTextEntry={placeholder === "Senha" ? true : false}
-        />
+          secureTextEntry={!showPassword && placeholder === "Senha"}
+          />
+          {["Senha", "Confirme sua Senha", "Nova senha"].includes(placeholder) && (
+            <Icon
+              name={showPassword ? "eye-off-outline" : "eye-outline"}
+              size={25}
+              color="black"
+              onPress={toggleShowPassword}
+              style={{ position: "absolute", right: 10, top: 11 }}
+            />
+          )}
       </ContainerLogin>
 
       {err !== "" && (