+++ src/screen/login/login.tsx
@@ -1,21 +1,16 @@
 import React, { useState } from "react";
 import { Platform, TouchableOpacity, Image } from "react-native";
 import { useNavigation } from "@react-navigation/native";
-
 import { propsStack } from "../../routes/Stack/Models";
-
 import {
   ButtonSenha,
   ContainerLogin,
   ContainerSafe, Scroll
 } from "./login-style";
-
 import { ContainerFormInputsPatients } from "../password-creation/cadastro-style";
-
 import logo from "../../assets/logo/logo.png";
 import logoEmail from "../../assets/logo/logoEmail.png";
 import logoSenha from "../../assets/logo/logoSenha.png";
-
 import Header from "../../components/Headers/header/header";
 import FooterCurto from "../../components/Footers/footer/footer";
 import FormPatientLoginRecord from "../../components/Forms/form-patient-login/form-patient-login";
@@ -36,8 +31,8 @@ const PatientLogin = () => {
   const [loading, setLoading] = useState(false);
 
   const arrayNamePlaceholder = [
-    [logoEmail, mail, "E-mail", errorMail],
-    [logoSenha, pass, "Senha", errorPass],
+    [ mail, "E-mail", errorMail],
+    [ pass, "Senha", errorPass],
   ];
 
   const handleInputChange = (name: string, value: string) => {
@@ -72,23 +67,20 @@ const PatientLogin = () => {
         <Header back={false} />
         <ContainerLogin>
           <Image source={logo} />
-
-          <ContainerFormInputsPatients>
+          <ContainerFormInputsPatients />
+            
             {arrayNamePlaceholder.map(
-              ([icon, valueState, place, err], index: number) => (
+              ([ valueState, place, err]) => (
                 <FormPatientLoginRecord
-                  imageButton={icon}
-                  placeholder={place}
-                  handleInputChange={(text: string) => {
-                    handleInputChange(place, text);
-                  }}
-                  state={valueState}
-                  err={err}
-                  key={index}
-                />
+                        placeholder={place}
+                        handleInputChange={(text: string) => {
+                            handleInputChange(place, text);
+                        } }
+                        state={valueState}
+                        err={err}
+                        showPassword={false}                />
               )
             )}
-          </ContainerFormInputsPatients>
 
           <CheckBox
             title="Lembrar senha"